refactor(nivona): stage 2 — extract brand-level primitives
Move five brand-wide concerns out of nivona/__init__.py into
private sibling modules:

- _crypto.py        — _NIVONA_RC4_KEY, _NIVONA_HU_TABLE
- _options.py       — setting-option enums (hardness, auto-off, profile, …)
- _registers.py     — RECIPE/TEMP/MYCOFFEE register bases + helpers
- _stats_helpers.py — _count / _pct / _flag factories
- _prefixes.py      — _MODEL_OVERRIDES, _MODEL_SETTINGS_EXCLUDE, _PREFIX_TO_FAMILY

__init__.py re-imports all of these so external callers (sensor.py,
_ble_commands.py, _ble_recipes.py, tests) keep their existing import
paths through `brands.nivona`. No behavior changes. 990 tests pass.
